requirements.txt
agate==1.6.3
agate-dbf==0.2.2
agate-excel==0.2.5
agate-sql==0.5.8
altair==5.0.1
annotated-types==0.5.0
anyio==3.7.1
attrs==19.3.0
Automat==0.8.0
Babel==2.10.3
backports.zoneinfo==0.2.1
blinker==1.4
cachetools==5.3.1
certifi==2019.11.28
chardet==3.0.4
Click==7.0
cloud-init==22.1
colorama==0.4.3
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==2.8
csvkit==1.0.7
dbfread==2.0.7
dbus-python==1.2.16
distro==1.4.0
distro-info===0.23ubuntu1
entrypoints==0.3
et-xmlfile==1.1.0
exceptiongroup==1.1.2
fastapi==0.101.0
future==0.18.2
gitdb==4.0.10
GitPython==3.1.32
greenlet==1.1.2
h11==0.14.0
httplib2==0.14.0
hyperlink==19.0.0
idna==2.8
imbalanced-learn==0.11.0
importlib-metadata==1.5.0
incremental==16.10.1
isodate==0.6.1
Jinja2==2.10.1
joblib==1.3.1
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
leather==0.3.4
markdown-it-py==3.0.0
MarkupSafe==1.1.0
mdurl==0.1.2
more-itertools==4.2.0
netifaces==0.10.4
numpy==1.24.4
oauthlib==3.1.0
olefile==0.46
openpyxl==3.0.10
packaging==23.1
pandas==2.0.3
parsedatetime==2.4
pexpect==4.6.0
Pillow==9.5.0
protobuf==4.24.0
pyarrow==12.0.1
pyasn1==0.4.2
pyasn1-modules==0.2.1
pydantic==2.1.1
pydantic-core==2.4.0
pydeck==0.8.0
Pygments==2.16.1
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==1.7.1
pymacaroons==0.13.0
Pympler==1.0.1
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyrsistent==0.15.5
pyserial==3.4
python-apt==2.0.0+ubuntu0.20.4.7
python-dateutil==2.8.2
python-debian===0.1.36ubuntu1
python-slugify==6.1.2
pytimeparse==1.1.8
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==5.3.1
requests==2.22.0
requests-unixsocket==0.2.0
rich==13.5.2
sample-stream==0.2.5
scikit-learn==1.3.0
scipy==1.10.1
SecretStorage==2.3.1
service-identity==18.1.0
simplejson==3.16.0
six==1.14.0
sklearn==0.0.post7
smmap==5.0.0
sniffio==1.3.0
sos==4.3
SQLAlchemy==1.4.39
ssh-import-id==5.10
starlette==0.27.0
streamlit==1.25.0
systemd-python==234
tenacity==8.2.2
text-unidecode==1.3
threadpoolctl==3.2.0
toml==0.10.2
toolz==0.12.0
tornado==6.3.3
Twisted==18.9.0
typing-extensions==4.7.1
tzdata==2023.3
tzlocal==4.3.1
ubuntu-advantage-tools==27.6
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
uvicorn==0.23.2
validators==0.21.2
wadllib==1.3.3
watchdog==3.0.0
xlrd==2.0.1
zipp==1.0.0
zope.interface==4.7.1
